Both setup pages complete
Finished up the setup local page and created setup bot page. Player names can be entered and will be displayed on game board now.

Problems: The choose move order has been bugged from before: randomize move button won't display on the radio input after a few presses, don't know why.

To be added in the future: The difficulty buttons stores the difficulty pressed, but there is no on screen indication as to which button is pressed. This should be added by someone in the future.